requirements.txt
# Core AI/ML
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
torch>=2.0.0

# Resume Parsing
pdfplumber>=0.10.0
python-docx>=1.0.0
PyPDF2>=3.0.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# UI
streamlit>=1.28.0

# Data & Utils
pandas>=2.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
